requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic-settings>=2.4.0
python-multipart>=0.0.9
aiofiles>=24.1.0

langchain>=0.2.14
langchain-community>=0.2.12
langchain-core>=0.2.34
langchain-groq>=0.1.9
langchain-huggingface>=0.0.3
langchain-text-splitters>=0.2.2
chromadb>=0.5.5
sentence-transformers>=3.0.1

pypdf>=4.3.1
python-docx>=1.1.2

httpx>=0.27.0
pytest>=8.3.2
pytest-asyncio>=0.23.8
ruff>=0.6.3
